third_party/WebKit/LayoutTests/media/video-seek-no-src.html

<!DOCTYPE HTML>
<title>Test that seeking video with no "src" attribute, sets default playback start position to that time.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
test(function() {
    var video = document.querySelector("video");
    assert_equals(video.networkState, HTMLMediaElement.NETWORK_EMPTY);
    assert_equals(video.readyState, HTMLMediaElement.HAVE_NOTHING);
    video.currentTime = 1.1;
    assert_equals(video.currentTime, 1.1);
});
</script>
